* | Dictionary | Update DictionaryKeyUInt

+++ Source/Dictionary/Keys/DictionaryKeyUInt.swift
@@ -9,13 +9,13 @@ public struct DictionaryKeyUInt: DictionaryKey {
     }
 
     public func serialize(src: any DictionaryKeyTypes) throws -> BigInt {
-        guard let src = src as? UInt32 else {
+        guard let src = src as? UInt64 else {
             throw TonError.custom("Key is not a uint")
         }
 
         return BigInt(
             try Builder()
-                .storeUint(UInt64(src), bits: bits)
+                .storeUint(src, bits: bits)
                 .endCell()
                 .beginParse()
                 .loadUintBig(bits: bits)
